Adjust ticks based on size
