+++ src/client/visualizations/scatterplot/scatterplot.tsx
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { Datum } from "plywood";
 import * as React from "react";
 
 import { ChartProps } from "../../../common/models/chart-props/chart-props";
@@ -23,26 +24,16 @@ import {
   VisualizationProps
 } from "../../views/cube-view/center-panel/center-panel";
 
-import * as d3 from "d3";
-import { Dataset, Datum } from "plywood";
-import { ConcreteSeries } from "../../../common/models/series/concrete-series";
-import { selectFirstSplitDatums } from "../../utils/dataset/selectors/selectors";
 import "./scatterplot.scss";
 
-import memoizeOne from "memoize-one";
-import { Essence } from "../../../common/models/essence/essence";
-import { Stage } from "../../../common/models/stage/stage";
 import { GridLines } from "../../components/grid-lines/grid-lines";
-import { LinearScale, pickTicks } from "../../utils/linear-scale/linear-scale";
 import { Point } from "./point";
 import { Tooltip } from "./tooltip";
+import { calculateXAxisStage, calculateYAxisStage, getPlottingData } from "./utils/get-plotting-data";
 import { XAxis } from "./x-axis";
 import { YAxis } from "./y-axis";
 
 const TICK_SIZE = 10;
-const MARGIN = 40;
-const X_AXIS_HEIGHT = 50;
-const Y_AXIS_WIDTH = 50;
 
 interface ScatterplotState {
   hoveredPoint: Datum | null;
@@ -62,16 +53,6 @@ export class Scatterplot extends React.Component<ChartProps, ScatterplotState> {
   render() {
     const { data, essence, stage } = this.props;
     const splitKey = essence.splits.splits.first().toKey();
-    // const [xSeries, ySeries] = essence.getConcreteSeries().toArray();
-    // const scatterplotData = selectFirstSplitDatums(data);
-    // const xExtent = getExtent(scatterplotData, xSeries);
-    // const yExtent = getExtent(scatterplotData, ySeries);
-    //
-    // const plottingStage = getPlottingStage(stage);
-    // const yScale = d3.scale.linear().domain(yExtent).nice().range([plottingStage.height, 0]);
-    // const xScale = d3.scale.linear().domain(xExtent).nice().range([0, plottingStage.width]);
-    // const xTicks = pickTicks(xScale, 10);
-    // const yTicks = pickTicks(yScale, 10);
 
     const { xTicks, yTicks, xScale, yScale, xSeries, ySeries, plottingStage, scatterplotData } = getPlottingData(data, essence, stage);
 
@@ -82,9 +63,9 @@ export class Scatterplot extends React.Component<ChartProps, ScatterplotState> {
       <svg viewBox={stage.getViewBox()}>
         <GridLines orientation={"vertical"} stage={plottingStage} ticks={xTicks} scale={xScale} />
         <GridLines orientation={"horizontal"} stage={plottingStage} ticks={yTicks} scale={yScale} />
-        <XAxis scale={xScale} stage={getXAxisStage(plottingStage)} ticks={xTicks} formatter={xSeries.formatter()} tickSize={TICK_SIZE}/>
+        <XAxis scale={xScale} stage={calculateXAxisStage(plottingStage)} ticks={xTicks} formatter={xSeries.formatter()} tickSize={TICK_SIZE}/>
         <YAxis
-          stage={getYAxisStage(plottingStage)}
+          stage={calculateYAxisStage(plottingStage)}
           ticks={yTicks}
           tickSize={TICK_SIZE}
           scale={yScale}
@@ -115,67 +96,3 @@ export function ScatterplotVisualization(props: VisualizationProps) {
     <ChartPanel {...props} queryFactory={makeQuery} chartComponent={Scatterplot}/>
   </React.Fragment>;
 }
-
-const getPlottingStage = memoizeOne(calculatePlottingStage);
-const getXAxisStage = memoizeOne(calculateXAxisStage);
-const getYAxisStage = memoizeOne(calculateYAxisStage);
-
-function getExtent(data: Datum[], series: ConcreteSeries): number[] {
-  const selectValues = (d: Datum) => series.selectValue(d);
-  return d3.extent(data, selectValues);
-}
-
-function calculatePlottingStage(stage: Stage): Stage {
-  return Stage.fromJS({
-    x: Y_AXIS_WIDTH + MARGIN,
-    y: MARGIN,
-    width: stage.width - Y_AXIS_WIDTH - 2 * MARGIN,
-    height: stage.height - X_AXIS_HEIGHT - 2 * MARGIN
-  });
-}
-
-function calculateXAxisStage(stage: Stage): Stage {
-  return Stage.fromJS({
-    x: Y_AXIS_WIDTH + MARGIN,
-    y: stage.height + MARGIN,
-    width: stage.width,
-    height: X_AXIS_HEIGHT
-  });
-}
-
-function calculateYAxisStage(stage: Stage): Stage {
-  return Stage.fromJS({
-    x: MARGIN,
-    y: MARGIN,
-    width: Y_AXIS_WIDTH,
-    height: stage.height
-  });
-}
-
-interface PlottingData {
-  xSeries: ConcreteSeries;
-  ySeries: ConcreteSeries;
-  xScale: LinearScale;
-  yScale: LinearScale;
-  xTicks: number[];
-  yTicks: number[];
-  plottingStage: Stage;
-  scatterplotData: Datum[];
-}
-
-function preparePlottingData(data: Dataset, essence: Essence, stage: Stage): PlottingData {
-  const [xSeries, ySeries] = essence.getConcreteSeries().toArray();
-  const scatterplotData = selectFirstSplitDatums(data);
-  const xExtent = getExtent(scatterplotData, xSeries);
-  const yExtent = getExtent(scatterplotData, ySeries);
-
-  const plottingStage = calculatePlottingStage(stage);
-  const yScale = d3.scale.linear().domain(yExtent).nice().range([plottingStage.height, 0]);
-  const xScale = d3.scale.linear().domain(xExtent).nice().range([0, plottingStage.width]);
-  const xTicks = pickTicks(xScale, 10);
-  const yTicks = pickTicks(yScale, 10);
-
-  return { xSeries, ySeries, xScale, yScale, xTicks, yTicks, plottingStage, scatterplotData };
-}
-
-const getPlottingData = memoizeOne(preparePlottingData);
